✨feat: change envars to writable ones for buildkite
Also adjusts the selector routine to use the repoPath instead.
